Add files via upload
This is movie recommendation in python created by Zubair Yousuf and Musadiq Shafique as AI project. This project contain recommendation of movies from 1996 to 2018 of Hollywood with the bases of Movies and rating. It contain both project documentation and sources file for python as well as jupyter. Me and Musadiq shafique both have work hard for this project so appreciation would be respected and its further use and download is open for every one for learning purpose Thank You.